feat(core-static-generator): :building_construction: add config.staticOut.targetBase feature
A first deployment test quickly determined that the static file generator would bake "localhost" into the base tag and some other, since that's where the script has been built. A config has been added that overrides the automatically detected base url in all files during file export. The key staticOut.targetBase must now be configured to match the intended url, to which the export is going to be deployed.