refactor(css): :children_crossing: reduce background visibility on short pages
Up until now, the lower half of the background showed on pages that were too short to reach the footer. That looked odd as soon as a background image was added. To avoid this issue, the content element (which covers the background) rather than the main element (which bears the background) has been chosen to extend to the full height on short pages.